fix some comments and move tests
